requirements.txt
aiohttp
python-socketio
python-socketio[client]
SQLAlchemy